+++ src/common_clj/component/datomic.clj
@@ -3,16 +3,16 @@
             [datomic.api :as d]
             [taoensso.timbre :as log]))
 
-(defn mocked-datomic [datomic-schemas]
+(defn ^:deprecated mocked-datomic [datomic-schemas]
   (let [datomic-uri "datomic:mem://unit-tests"
         connection (do (d/create-database datomic-uri)
                        (d/connect datomic-uri))]
     @(d/transact connection (flatten datomic-schemas))
     connection))
 
-(defrecord Datomic [config schemas]
+(defrecord ^:deprecated Datomic [config schemas]
   component/Lifecycle
-  (start [component]
+  (start ^:deprecated [component]
     (let [datomic-uri (or (-> config :config :datomic-uri)
                           "datomic:mem://integration-tests")
           connection (do (log/info ::database-creation (d/create-database datomic-uri))
@@ -20,9 +20,9 @@
       @(d/transact connection (flatten schemas))
       (assoc component :datomic connection)))
 
-  (stop [{:keys [datomic] :as component}]
+  (stop ^:deprecated [{:keys [datomic] :as component}]
     (d/release datomic)
     (assoc component :datomic nil)))
 
-(defn new-datomic [schemas]
+(defn ^:deprecated new-datomic [schemas]
   (map->Datomic {:schemas schemas}))