+++ src/common_clj/test/helper/auth.clj
@@ -1,1 +1,0 @@
-(ns common-clj.test.helper.auth)